Conflicts:
	host/lib/usrp/usrp2/codec_impl.cpp

+++ host/include/uhd/usrp/codec_props.hpp
@@ -1,5 +1,5 @@
 //
-// Copyright 2010 Ettus Research LLC
+// Copyright 2010-2011 Ettus Research LLC
 //
 // This program is free software: you can redistribute it and/or modify
 // it under the terms of the GNU General Public License as published by
@@ -30,8 +30,8 @@ namespace uhd{ namespace usrp{
     enum codec_prop_t{
         CODEC_PROP_NAME           = 'n', //ro, std::string
         CODEC_PROP_OTHERS         = 'o', //ro, prop_names_t
-        CODEC_PROP_GAIN_I         = 'i', //rw, float
-        CODEC_PROP_GAIN_Q         = 'q', //rw, float
+        CODEC_PROP_GAIN_I         = 'i', //rw, double
+        CODEC_PROP_GAIN_Q         = 'q', //rw, double
         CODEC_PROP_GAIN_RANGE     = 'r', //ro, gain_range_t
         CODEC_PROP_GAIN_NAMES     = 'G'  //ro, prop_names_t
     };
